Move scenarios.ts into providers.ts
